requirements.txt
dash==2.18.2
dash-bootstrap-components==1.7.1
plotly==5.24.1
flask==3.0.3
gunicorn==23.0.0
sqlalchemy==2.0.36
pandas==2.2.3
pyyaml==6.0.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pymysql==1.1.1
pyodbc==5.0.1
werkzeug==3.0.4
